personal-knowledge-assistant/requirements.txt
endee>=0.1.0
sentence-transformers>=2.7.0
pdfplumber>=0.11.0
google-genai>=0.8.0
openai>=1.0.0
streamlit>=1.35.0
python-dotenv>=1.0.0
rich>=13.0.0
